duckdb_query_to_parquet_src/requirements.txt
functions-framework==3.3.0
requests==2.28.2
google-cloud-storage==2.8.0
duckdb==0.7.1
google-cloud-bigquery==3.9.0
